requirements.txt
# project/requirements.txt
torch
torchvision
timm
optuna
mlflow
pyyaml
tqdm
numpy
